backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.1
python-dotenv==1.0.0
httpx==0.27.0
beautifulsoup4==4.12.3
scrapy==2.11.0